extract background responses tests into a separate file
